requirements.txt
grpcio
grpcio-tools
requests
pyyaml